Create new Kirby3 Repo, Move to kirbyup and detach from K2 version.
