+++ index.css
@@ -1,91 +1,1 @@
-.k-button-group > .k-button.translations--active, .k-button-group > .k-dropdown.translations--active {
-  background-color: rgba(255,255,255,.5);
-}
-
-/* Clear kirby styles, looks better */
-.k-button-group .k-translations-options.k-dropdown-content {
-  margin: 0;
-  padding: 0 var(--padding-x);
-}
-
-/* inject style to hide the kirby lang menu when portaled */
-.k-translations-hide-kirbylangs .k-header-buttons .k-languages-dropdown {
-  display: none;
-}
-.k-translations-hide-kirbylangs .k-header-buttons .k-languages-dropdown.k-translations-languages-dropdown {
-  display: inline-block;
-}
-
-/* Don't display the section when portaled */
-.k-section.k-translations-field.portaled {
-  display: none;
-  padding-bottom: 0;
-}
-
-/* Our buttons always have text but their icons can be hidden (overrides some Kirby css too) */
-.k-translations-button.k-button[data-responsive] .k-button-text {
-  display: inline;
-}
-.k-translations-button .k-button-text {
-  display: inline;
-}
-.k-translations-button .k-button-text .shortname {
-  display: inline;
-  text-transform: capitalize;
-}
-.k-translations-button .k-button-text .longname {
-  display: none;
-}
-[dir=ltr] .k-translations-button .k-button-icon~.k-button-text {
-  padding-left: 0;
-}
-.k-translations-button .k-icon {
-  display: none;
-}
-.k-translations-menu-icon {
-  color: var(--color-gray-900);
-}
-
-/* In portal mode, adding more languages to the headerbar breaks the layout. Ensure our buttons remain small. */
-@media screen and (min-width: 50em) {
-  .k-button[data-responsive] .k-button-text .shortname {
-    display: none;
-  }
-  .k-button[data-responsive] .k-button-text .longname {
-    display: inline-block;
-  }
-}
-/* Kirby's default responsive rulez from 30 to 35em */
-@media screen and (min-width: 30em) {
-  .k-translations-hide-kirbylangs .k-header .k-button[data-responsive]:not(.k-translations-button) .k-button-text {
-    display: none;
-  }
-  .k-translations-button.k-button[data-responsive] .k-button-text {
-    display: inline;
-  }
-}
-@media screen and (min-width: 35em) {
-  .k-translations-hide-kirbylangs .k-header .k-button[data-responsive]:not(.k-translations-button) .k-button-text {
-    display: inline-block;
-  }
-}
-@media screen and (min-width: 40em) {
-  .k-translations-button .k-icon {
-    display: inherit;
-  }
-  .k-translations-button .k-button-text {
-    display: inherit;
-  }
-  [dir=ltr] .k-translations-button .k-button-icon~.k-button-text {
-    padding-left: 0.5rem;
-  }
-}
-
-/*.k-button-group > .k-button.translations--active::after {
-  content: '';
-  display: block;
-  width: 100%;
-  height: 2px;
-  background-color: #999;
-  margin-top: .5rem;
-}*/
+.k-button-group>.k-button.translations--active,.k-button-group>.k-dropdown.translations--active{background-color:#ffffff80}.k-button-group .k-translations-options.k-dropdown-content{margin:0;padding:0 var(--padding-x)}.k-translations-hide-kirbylangs .k-header-buttons .k-languages-dropdown{display:none}.k-translations-hide-kirbylangs .k-header-buttons .k-languages-dropdown.k-translations-languages-dropdown{display:inline-block}.k-section.k-translations-field.portaled{display:none;padding-bottom:0}.k-translations-button.k-button[data-responsive] .k-button-text,.k-translations-button .k-button-text{display:inline}.k-translations-button .k-button-text .shortname{display:inline;text-transform:capitalize}.k-translations-button .k-button-text .longname{display:none}[dir=ltr] .k-translations-button .k-button-icon~.k-button-text{padding-left:0}.k-translations-button .k-icon{display:none}.k-translations-menu-icon{color:var(--color-gray-900)}@media screen and (min-width: 50em){.k-button[data-responsive] .k-button-text .shortname{display:none}.k-button[data-responsive] .k-button-text .longname{display:inline-block}}@media screen and (min-width: 30em){.k-translations-hide-kirbylangs .k-header .k-button[data-responsive]:not(.k-translations-button) .k-button-text{display:none}.k-translations-button.k-button[data-responsive] .k-button-text{display:inline}}@media screen and (min-width: 35em){.k-translations-hide-kirbylangs .k-header .k-button[data-responsive]:not(.k-translations-button) .k-button-text{display:inline-block}}@media screen and (min-width: 40em){.k-translations-button .k-icon,.k-translations-button .k-button-text{display:inherit}[dir=ltr] .k-translations-button .k-button-icon~.k-button-text{padding-left:.5rem}}
