+++ index.js
@@ -1,241 +1,1 @@
-panel.plugin('flokosiol/translations', {
-  sections: {
-    translations: {
-      data: function () {
-        return {
-          id: null,
-          deletable: null,
-          revertable: null,
-          translations: null,
-          compactmode: null,
-          portaled: null
-        }
-      },
-      computed: {
-        // Computed functions very similar to /panel/src/components/Navigation/Languages.vue
-        defaultLanguage() {
-          // Kirby 3.6 Fiber
-          if (this.hasFiber) {
-            return window.panel.$languages.find(l => l.default == true) ?? window.panel.$languages[0];
-          }
-          // Pre 3.6
-          return this.$store.state.languages.default;
-        },
-        language() {
-          if (this.hasFiber) {
-            return window.panel.$language;
-          }
-          return this.$store.state.languages.current;
-        },
-        languages() {
-          if (this.hasFiber) {
-            return window.panel.$languages.filter(language => language.default === false);
-          }
-          return this.$store.state.languages.all.filter(language => language.default === false);
-        },
-        hasFiber() {
-          return window.panel && window.panel.$languages;
-        },
-        portalSelector() {
-          return '.k-header-buttons div.k-bar-slot[data-position=left] .k-button-group';
-        },
-        compactModeEnabled() {
-          return this.compactmode || this.portaled;
-        }
-      },
-      created: function() {
-        this.updateButtons();
-        this.$events.$on("model.save", this.updateButtons);
-        this.$events.$on("model.update", this.updateButtons);
-      },
-      methods: {
-        updateButtons() {
-          this.load().then(response => {
-            // populate the data object when the section gets loaded
-            // @see https://getkirby.com/docs/reference/plugins/extensions/sections#vue-component
-            this.id           = response.id;
-            this.deletable    = response.deletable;
-            this.revertable   = response.revertable;
-            this.translations = response.translations;
-            this.compactmode  = response.compactmode;
-            this.updatePortaled(response.portaled);
-
-            // loop through all panel languages (but default)
-            let languages = this.languages;
-            for (let i = 0; i < languages.length; i++) {
-              // set icon + color if page is translated
-              if (this.translations.indexOf(languages[i].code) >= 0) {
-                this.languages[i].icon = 'check';
-                this.languages[i].theme = 'positive';
-              }
-              else {
-                this.languages[i].icon = 'cancel';
-                this.languages[i].theme = 'negative';
-              }
-            }
-
-            // hide actions (delete + revert) for default language + untranslated languages
-            if (this.language.code === this.defaultLanguage.code || this.translations.indexOf(this.language.code) < 0) {
-              this.deletable = false;
-              this.revertable = false;
-            }
-          });
-        },
-        changeLanguage(language) {
-          this.updateButtons();
-
-          // execute language change
-          if (this.hasFiber) {
-            this.$emit("change", language);
-            this.$go(this.$view.path, {
-              query: {
-                language: language.code
-              }
-            });
-          }
-          else {
-            this.$store.dispatch("languages/current", language);
-            this.$emit("change", language);
-          }
-        },
-        deleteTranslationOpen(id, language) {
-          this.$refs.deleteDialog.open(id, language);
-        },
-        deleteTranslationSubmit(id, language) {
-          this.$api.post('flokosiol/translations/delete', {id: id, languageCode: language.code})
-            .then(response => {
-              this.$refs.deleteDialog.close();
-              if (response.code === 200) {
-                this.$store.dispatch('notification/success', response.text);
-                this.changeLanguage(this.defaultLanguage);
-              }
-              else {
-                this.$store.dispatch('notification/error', response.text);
-              }
-            })
-            .catch(error => {
-              this.$store.dispatch('notification/error', error);
-            });
-        },
-        revertTranslationOpen(id, language) {
-          this.$refs.revertDialog.open(id, language);
-        },
-        revertTranslationSubmit(id, language) {
-          this.$api.post('flokosiol/translations/revert', {id: id, languageCode: language.code})
-            .then(response => {
-              this.$refs.revertDialog.close();
-              if (response.code === 200) {
-                this.$store.dispatch('notification/success', response.text);
-                this.updateButtons();
-
-                if (this.hasFiber) this.$go(this.$view.path);
-              }
-              else {
-                this.$store.dispatch('notification/error', response.text);
-              }
-            })
-            .catch(error => {
-              this.$store.dispatch('notification/error', error);
-            });
-
-          // reload panel to read changed textfile
-          if(!this.hasFiber) this.$router.go();
-        },
-        updatePortaled(value) {
-          if(value) this.portaled = value;
-
-          // Update portaled state to hide panel lang switcher via css
-          this.portaled ? document.body.classList.add('k-translations-hide-kirbylangs') : document.body.classList.remove('k-translations-hide-kirbylangs');
-        },
-        menuLanguageClick(language) {
-          // Clicking on the current lang toggles the submenu
-          if (language.code == this.language.code) {
-            if ( this.compactModeEnabled ) {
-              if (this.getLanguageOptionsToggleRef()) this.getLanguageOptionsToggleRef().toggle();
-            }
-            else {
-              // ignore clicks on current language (triggers a useless reload)
-              // this.changeLanguage(language); to reload page to replicate original behaviour
-            }
-          }
-          else this.changeLanguage(language);
-        },
-        getLanguageOptionsToggleRef() {
-          // Sometimes when 2 items fill the same $ref, it becomes an array...
-          if(this.$refs.languageOptionsToggle){
-            return (this.$refs.languageOptionsToggle instanceof Array) ? this.$refs.languageOptionsToggle.at(0) : this.$refs.languageOptionsToggle;
-          }
-          return null;
-        },
-        getLangButtonClass(lang, extra = {}) {
-          return {
-            'translations--active': lang.code === this.language.code, // mark active
-            'k-translations-button': true, // custom styles
-            ...extra
-          };
-        },
-      },
-      template: `
-        <div :class="{'portaled': portaled, 'k-field':true, 'k-translations-field':true}">
-          
-          <k-button-group v-if="languages.length" class="k-translations-buttons">
-            <portal :selector="portalSelector" :disabled="!portaled">
-              <k-button :data-langcode="defaultLanguage.code" icon="check" :class="getLangButtonClass(defaultLanguage, {'k-languages-dropdown': true, 'k-translations-languages-dropdown': true})" theme="positive" :key="defaultLanguage.code" @click.stop="menuLanguageClick(defaultLanguage)" :responsive="compactModeEnabled">
-                  <span :class="{'longname':compactModeEnabled}">{{ defaultLanguage.name }}</span>
-                  <span class="shortname" v-if="compactModeEnabled">{{ defaultLanguage.code }}</span>
-              </k-button>
-            </portal>
-            <portal :selector="portalSelector" v-for="lang in languages" :key="lang.code" :disabled="!portaled">
-                <k-dropdown :class="getLangButtonClass(lang, {'k-languages-dropdown': true, 'k-translations-languages-dropdown': true})">
-                  <k-button class="k-translations-button" :icon="lang.icon" :theme="lang.theme" @click.stop="menuLanguageClick(lang)" :responsive="compactModeEnabled">
-                    <span :class="{'longname':compactModeEnabled}">{{ lang.name }}</span>
-                    <span v-if="compactModeEnabled" class="shortname">{{ lang.code }}</span>
-                    <k-icon v-if="compactModeEnabled && language.code == lang.code" class="k-translations-menu-icon" type="angle-down" />
-                  </k-button>
-                  <k-dropdown-content v-if="compactModeEnabled && (lang.code === language.code)" class="k-translations-options" ref="languageOptionsToggle">
-                    <k-button-group>
-                      <k-dropdown-item v-if="deletable" icon="trash" @click.stop="deleteTranslationOpen(id, language)">
-                        {{ $t('delete') }} {{ language.code.toUpperCase() }}
-                      </k-dropdown-item>
-                      <k-dropdown-item v-if="revertable" icon="refresh" @click.stop="revertTranslationOpen(id, language)">
-                        {{ $t('revert') }} {{ language.code.toUpperCase() }}
-                      </k-dropdown-item>
-                    </k-button-group>
-                  </k-dropdown-content>
-                </k-dropdown>
-              </portal>
-          </k-button-group>
-          
-          <k-button-group v-if="!compactModeEnabled" class="k-translations-options">
-            <k-button v-if="deletable" icon="trash" @click.stop="deleteTranslationOpen(id, language)">
-              {{ $t('delete') }} {{ language.code.toUpperCase() }}
-            </k-button>
-            <k-button v-if="revertable" icon="refresh" @click.stop="revertTranslationOpen(id, language)">
-              {{ $t('revert') }} {{ language.code.toUpperCase() }}
-            </k-button>
-          </k-button-group>
-
-          <k-dialog
-            ref="deleteDialog"
-            :button="$t('delete')"
-            theme="negative"
-            icon="trash"
-            @submit="deleteTranslationSubmit(id, language)"
-          >
-            <k-text v-html="$t('flokosiol.translations.delete.confirm', { code: language.code.toUpperCase() })" />
-          </k-dialog>
-
-          <k-dialog
-            ref="revertDialog"
-            :button="$t('revert')"
-            theme="negative"
-            icon="refresh"
-            @submit="revertTranslationSubmit(id, language)"
-          >
-            <k-text v-html="$t('flokosiol.translations.revert.confirm', { code: language.code.toUpperCase() })" />
-          </k-dialog>
-        </div>
-      `
-    }
-  }
-});
+var M=Object.defineProperty;var m=Object.getOwnPropertySymbols;var E=Object.prototype.hasOwnProperty,y=Object.prototype.propertyIsEnumerable;var v=(l,s,i)=>s in l?M(l,s,{enumerable:!0,configurable:!0,writable:!0,value:i}):l[s]=i,b=(l,s)=>{for(var i in s||(s={}))E.call(s,i)&&v(l,i,s[i]);if(m)for(var i of m(s))y.call(s,i)&&v(l,i,s[i]);return l};(function(){"use strict";var l=function(){var e=this,a=e.$createElement,t=e._self._c||a;return t("div",{class:{portaled:e.portaled,"k-field":!0,"k-translations-field":!0}},[e.languages.length?t("k-button-group",{staticClass:"k-translations-buttons"},[t("portal",{attrs:{selector:e.portalSelector,disabled:!e.portaled}},[t("k-button",{key:e.defaultLanguage.code,class:e.getLangButtonClass(e.defaultLanguage,{"k-languages-dropdown":!0,"k-translations-languages-dropdown":!0}),attrs:{"data-langcode":e.defaultLanguage.code,icon:"check",theme:"positive",responsive:e.compactModeEnabled},on:{click:function(n){return n.stopPropagation(),e.menuLanguageClick(e.defaultLanguage)}}},[t("span",{class:{longname:e.compactModeEnabled}},[e._v(e._s(e.defaultLanguage.name))]),e.compactModeEnabled?t("span",{staticClass:"shortname"},[e._v(e._s(e.defaultLanguage.code))]):e._e()])],1),e._l(e.languages,function(n){return t("portal",{key:n.code,attrs:{selector:e.portalSelector,disabled:!e.portaled}},[t("k-dropdown",{class:e.getLangButtonClass(n,{"k-languages-dropdown":!0,"k-translations-languages-dropdown":!0})},[t("k-button",{staticClass:"k-translations-button",attrs:{icon:n.icon,theme:n.theme,responsive:e.compactModeEnabled},on:{click:function(r){return r.stopPropagation(),e.menuLanguageClick(n)}}},[t("span",{class:{longname:e.compactModeEnabled}},[e._v(e._s(n.name))]),e.compactModeEnabled?t("span",{staticClass:"shortname"},[e._v(e._s(n.code))]):e._e(),e.compactModeEnabled&&e.language.code==n.code?t("k-icon",{staticClass:"k-translations-menu-icon",attrs:{type:"angle-down"}}):e._e()],1),e.compactModeEnabled&&n.code===e.language.code?t("k-dropdown-content",{ref:"languageOptionsToggle",refInFor:!0,staticClass:"k-translations-options"},[t("k-button-group",[e.deletable?t("k-dropdown-item",{attrs:{icon:"trash"},on:{click:function(r){return r.stopPropagation(),e.deleteTranslationOpen(e.id,e.language)}}},[e._v(" "+e._s(e.$t("delete"))+" "+e._s(e.language.code.toUpperCase())+" ")]):e._e(),e.revertable?t("k-dropdown-item",{attrs:{icon:"refresh"},on:{click:function(r){return r.stopPropagation(),e.revertTranslationOpen(e.id,e.language)}}},[e._v(" "+e._s(e.$t("revert"))+" "+e._s(e.language.code.toUpperCase())+" ")]):e._e()],1)],1):e._e()],1)],1)})],2):e._e(),e.compactModeEnabled?e._e():t("k-button-group",{staticClass:"k-translations-options"},[e.deletable?t("k-button",{attrs:{icon:"trash"},on:{click:function(n){return n.stopPropagation(),e.deleteTranslationOpen(e.id,e.language)}}},[e._v(" "+e._s(e.$t("delete"))+" "+e._s(e.language.code.toUpperCase())+" ")]):e._e(),e.revertable?t("k-button",{attrs:{icon:"refresh"},on:{click:function(n){return n.stopPropagation(),e.revertTranslationOpen(e.id,e.language)}}},[e._v(" "+e._s(e.$t("revert"))+" "+e._s(e.language.code.toUpperCase())+" ")]):e._e()],1),t("k-dialog",{ref:"deleteDialog",attrs:{button:e.$t("delete"),theme:"negative",icon:"trash"},on:{submit:function(n){return e.deleteTranslationSubmit(e.id,e.language)}}},[t("k-text",{domProps:{innerHTML:e._s(e.$t("daandelange.translations.delete.confirm",{code:e.language.code.toUpperCase()}))}})],1),t("k-dialog",{ref:"revertDialog",attrs:{button:e.$t("revert"),theme:"negative",icon:"refresh"},on:{submit:function(n){return e.revertTranslationSubmit(e.id,e.language)}}},[t("k-text",{domProps:{innerHTML:e._s(e.$t("daandelange.translations.revert.confirm",{code:e.language.code.toUpperCase()}))}})],1)],1)},s=[],i="";function $(e,a,t,n,r,h,p,L){var o=typeof e=="function"?e.options:e;a&&(o.render=a,o.staticRenderFns=t,o._compiled=!0),n&&(o.functional=!0),h&&(o._scopeId="data-v-"+h);var u;if(p?(u=function(d){d=d||this.$vnode&&this.$vnode.ssrContext||this.parent&&this.parent.$vnode&&this.parent.$vnode.ssrContext,!d&&typeof __VUE_SSR_CONTEXT__!="undefined"&&(d=__VUE_SSR_CONTEXT__),r&&r.call(this,d),d&&d._registeredComponents&&d._registeredComponents.add(p)},o._ssrRegister=u):r&&(u=L?function(){r.call(this,(o.functional?this.parent:this).$root.$options.shadowRoot)}:r),u)if(o.functional){o._injectStyles=u;var w=o.render;o.render=function(O,_){return u.call(_),w(O,_)}}else{var f=o.beforeCreate;o.beforeCreate=f?[].concat(f,u):[u]}return{exports:e,options:o}}const k={components:{},data:function(){return{id:null,deletable:null,revertable:null,translations:null,compactmode:null,portaled:null}},computed:{defaultLanguage(){var e;return this.hasFiber?(e=window.panel.$languages.find(a=>a.default==!0))!=null?e:window.panel.$languages[0]:this.$store.state.languages.default},language(){return this.hasFiber?window.panel.$language:this.$store.state.languages.current},languages(){return this.hasFiber?window.panel.$languages.filter(e=>e.default===!1):this.$store.state.languages.all.filter(e=>e.default===!1)},hasFiber(){return window.panel&&window.panel.$languages},portalSelector(){return".k-header-buttons div.k-bar-slot[data-position=left] .k-button-group"},compactModeEnabled(){return this.compactmode||this.portaled}},created:function(){this.updateButtons(),this.$events.$on("model.save",this.updateButtons),this.$events.$on("model.update",this.updateButtons)},methods:{updateButtons(){this.load().then(e=>{this.id=e.id,this.deletable=e.deletable,this.revertable=e.revertable,this.translations=e.translations,this.compactmode=e.compactmode,this.updatePortaled(e.portaled);let a=this.languages;for(let t=0;t<a.length;t++)this.translations.indexOf(a[t].code)>=0?(this.languages[t].icon="check",this.languages[t].theme="positive"):(this.languages[t].icon="cancel",this.languages[t].theme="negative");(this.language.code===this.defaultLanguage.code||this.translations.indexOf(this.language.code)<0)&&(this.deletable=!1,this.revertable=!1)})},changeLanguage(e){this.updateButtons(),this.hasFiber?(this.$emit("change",e),this.$go(this.$view.path,{query:{language:e.code}})):(this.$store.dispatch("languages/current",e),this.$emit("change",e))},deleteTranslationOpen(e,a){this.$refs.deleteDialog.open(e,a)},deleteTranslationSubmit(e,a){this.$api.post("daandelange/translations/delete",{id:e,languageCode:a.code}).then(t=>{this.$refs.deleteDialog.close(),t.code===200?(this.$store.dispatch("notification/success",t.text),this.changeLanguage(this.defaultLanguage)):this.$store.dispatch("notification/error",t.text)}).catch(t=>{this.$store.dispatch("notification/error",t)})},revertTranslationOpen(e,a){this.$refs.revertDialog.open(e,a)},revertTranslationSubmit(e,a){this.$api.post("daandelange/translations/revert",{id:e,languageCode:a.code}).then(t=>{this.$refs.revertDialog.close(),t.code===200?(this.$store.dispatch("notification/success",t.text),this.updateButtons(),this.hasFiber&&this.$go(this.$view.path)):this.$store.dispatch("notification/error",t.text)}).catch(t=>{this.$store.dispatch("notification/error",t)}),this.hasFiber||this.$router.go()},updatePortaled(e){e&&(this.portaled=e),this.portaled?document.body.classList.add("k-translations-hide-kirbylangs"):document.body.classList.remove("k-translations-hide-kirbylangs")},menuLanguageClick(e){e.code==this.language.code?this.compactModeEnabled&&this.getLanguageOptionsToggleRef()&&this.getLanguageOptionsToggleRef().toggle():this.changeLanguage(e)},getLanguageOptionsToggleRef(){return this.$refs.languageOptionsToggle?this.$refs.languageOptionsToggle instanceof Array?this.$refs.languageOptionsToggle.at(0):this.$refs.languageOptionsToggle:null},getLangButtonClass(e,a={}){return b({"translations--active":e.code===this.language.code,"k-translations-button":!0},a)}}},g={};var C=$(k,l,s,!1,T,null,null,null);function T(e){for(let a in g)this[a]=g[a]}var c=function(){return C.exports}();panel.plugin("daandelange/translations",{components:{TranslationsSection:c},sections:{translations:c}})})();
