+++ src/index.js
@@ -1,0 +1,10 @@
+import TranslationsSection from "./components/TranslationsSection.vue";
+
+panel.plugin('daandelange/translations', {
+  components: {
+    TranslationsSection,
+  },
+  sections: {
+    translations: TranslationsSection,
+  },
+});
